# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 77Photo API
  version: 1.0.0
  description: |
    V1 REST API for the filesystem-first 77Photo family photo service.
    All application endpoints use JSON unless stated otherwise. Timestamps are
    RFC 3339 UTC strings. The API never accepts a server filesystem path from a
    client; resources are addressed by opaque IDs. Authenticated application
    endpoints accept either cookieAuth or bearerAuth. Cookie-authenticated
    writes require X-CSRF-Token; Bearer clients must not send browser CSRF
    tokens. Mobile login and refresh return tokens in JSON and never set
    browser cookies.
servers:
  - url: /
    description: Same-origin deployment
...
paths:
...
  /api/v1/setup/admin:
    post:
      tags: [setup]
      summary: Create the first administrator
      description: |
        The only unauthenticated write endpoint. It is available exactly once,
        while the database contains no users, and is intended to be called from
        a local trusted network during first startup. It returns 409 after the
        first user has been created.
      operationId: initializeAdmin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupAdminRequest' }
      responses:
        '201':
          description: Administrator created and a session established
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: Initialization has already completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Username or password does not satisfy policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/setup/status:
    get:
      tags: [setup]
      summary: Check whether first-run setup is required
      description: Returns only whether the installation has no users and still requires its first administrator.
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Current setup requirement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupStatusResponse' }
        '500':
          description: Setup state could not be determined
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/users:
    get:
      tags: [users]
      summary: List users
      operationId: listUsers
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      description: Includes deleted (tombstoned) accounts with deleted_at set; clients normally hide them.
      responses:
        '200':
          description: Users without password hashes or session secrets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminRequired' }
    post:
      tags: [users]
      summary: Create a user
      operationId: createUser
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminOrCSRFForbidden' }
        '409':
          description: Username already belongs to an existing account (USERNAME_TAKEN). Usernames of deleted accounts are free for reuse.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '422': { $ref: '#/components/responses/InvalidCredentials' }
...
  /api/v1/folders:
    get:
      tags: [folders]
      summary: List folders visible to the current user
      operationId: listFolders
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ParentFolderID'
      responses:
        '200':
          description: Direct children of the selected root
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Folder' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [folders]
      summary: Create a folder
      operationId: createFolder
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFolderRequest' }
      responses:
        '201':
          description: Folder created
          content: { application/json: { schema: { $ref: '#/components/schemas/Folder' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/WriteForbidden' }
        '409':
          description: A sibling folder has the same name
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/photos:
    get:
      tags: [photos]
      summary: List photos with stable cursor pagination
      operationId: listPhotos
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/FolderFilter'
        - name: q
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
          description: Trimmed filename substring, 100 Unicode characters maximum. ASCII letters match without case; non-ASCII characters match exactly. Percent and underscore are literal characters. Empty or whitespace-only values disable this filter.
        - name: media_type
          in: query
          required: false
          schema: { type: string, enum: [photo, video] }
          description: Photo includes still and Live Photos; video includes standalone videos. Omit for all media.
        - name: favorite
          in: query
          required: false
          schema: { type: string, enum: ['true'] }
          description: Return only the caller's currently visible favorites. Bound into the signed cursor and combinable with other filters.
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/BBoxFilter'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Photos sorted by captured_at DESC then id DESC
          content: { application/json: { schema: { $ref: '#/components/schemas/PhotoPage' } } }
        '400':
          description: Invalid timestamp, bounding box, cursor, or incompatible filters
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ReadForbidden' }
        '410':
          description: Cursor has expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: CURSOR_EXPIRED
                  message: cursor expired; restart pagination without cursor
                  request_id: req_01JABC
...
  /api/v1/photos/upload:
    post:
      tags: [photos]
      summary: Upload one photo or video
      description: |
        Each request contains exactly one file. Clients performing batch upload
        send independent requests and may limit concurrency. The server streams
        the body to a temporary file, validates content, fsyncs and atomically
        renames it before creating the index row.
      operationId: uploadPhoto
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, folder_id]
              properties:
                file:
                  type: string
                  format: binary
                folder_id:
                  $ref: '#/components/schemas/ID'
                conflict:
                  type: string
                  enum: [reject, rename]
                  default: reject
                file_modified_at:
                  type: string
                  format: date-time
                  description: Optional client-side original file modification time used only when embedded capture metadata is unavailable
      responses:
        '201':
          description: File stored and indexed
          content: { application/json: { schema: { $ref: '#/components/schemas/Photo' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/WriteForbidden' }
        '409':
          description: Duplicate content or name conflict
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '413':
          description: File exceeds PHOTO_MAX_UPLOAD_SIZE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: UPLOAD_TOO_LARGE
                  message: upload exceeds the configured maximum size
                  request_id: req_01JABC
        '415':
          description: Unsupported or content-type-mismatched media
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '422':
          description: Invalid or unsafe media (including excessive pixel count)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '507':
          description: Server storage or quota is full
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: STORAGE_FULL
                  message: server storage is full
                  request_id: req_01JABC
...
  /api/v1/me/share-links:
    get:
      tags: [share-links]
      summary: List share links manageable by the current user
      description: Current resource ownership determines access. Administrators also see orphan records. Tokens and password hashes are never returned.
      operationId: listManagedShareLinks
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, expired, revoked, unavailable] } }
        - { name: resource_type, in: query, schema: { type: string, enum: [photo, folder] } }
        - { name: resource_id, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string }, description: Opaque page cursor }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
      responses:
        '200':
          description: Share links sorted by creation time, newest first
          content: { application/json: { schema: { $ref: '#/components/schemas/ManagedSharePage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: Invalid filter or cursor, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /api/v1/share-links:
    post:
      tags: [share-links]
      summary: Create a read-only public link for a photo or folder
      description: |
        The caller must own the resource or be an administrator. The returned
        URL is a hash route such as /#/share/{token}; the token is never stored
        in plaintext by the server. A password is optional. Duration is one of
        1_day, 7_days, or forever.
      operationId: createShareLink
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateShareLinkRequest' }
      responses:
        '201':
          description: Public link created; recipients can only view scoped previews
          content: { application/json: { schema: { $ref: '#/components/schemas/ShareLink' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ShareManageForbidden' }
        '422':
          description: Resource, duration, or optional password is invalid
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/shares:
    get:
      tags: [shares]
      summary: Legacy member ACL shares (no longer shown in the Web UI)
      operationId: listShares
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Share records without secrets
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Share' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [shares]
      summary: Legacy member ACL share (no longer shown in the Web UI)
      operationId: createShare
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateShareRequest' } } }
      responses:
        '201':
          description: Share created
          content: { application/json: { schema: { $ref: '#/components/schemas/Share' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/ShareManageForbidden' }
        '409':
          description: Same folder is already shared with the user
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/admin/rescan:
    post:
      tags: [admin]
      summary: Start a filesystem rescan
      operationId: startRescan
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      responses:
        '202':
          description: Scan queued
          content: { application/json: { schema: { $ref: '#/components/schemas/RescanJob' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminOrCSRFForbidden' }
        '409':
          description: A rescan is already queued or running (RESCAN_IN_PROGRESS, details.job_id), or another maintenance task holds the lock (MAINTENANCE_IN_PROGRESS)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/admin/maintenance:
    get:
      tags: [admin]
      summary: Get the maintenance task currently holding the lock
      description: |
        Rescan/reset, thumbnail rebuild, import, broken photo cleanup and user
        photo transfer share one server-side lock, so at most one runs at a
        time. Clients use this endpoint to re-attach to a running job after a
        reload (job_id is set for rescan, thumbnail_rebuild and import) and
        after receiving MAINTENANCE_IN_PROGRESS.
      operationId: getMaintenance
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Active maintenance task, or null when idle
          content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStatus' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminRequired' }
  /api/v1/admin/thumbnails/rebuild:
    post:
      tags: [admin]
      summary: Start a thumbnail rebuild
      description: |
        incremental only renders photos with a missing variant. full re-renders
        every supported photo and video, replacing variants in place so a
        failed render keeps the existing thumbnails, and prunes cached variants
        of older source revisions. Original media is never modified.
      operationId: startThumbnailRebuild
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ThumbnailRebuildRequest' }
      responses:
        '202':
          description: Rebuild queued
          content: { application/json: { schema: { $ref: '#/components/schemas/ThumbnailRebuildJob' } } }
        '400':
          description: Invalid JSON (INVALID_REQUEST) or unknown mode (INVALID_REBUILD_MODE)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminOrCSRFForbidden' }
        '409':
          description: A rebuild is already running (THUMBNAIL_REBUILD_IN_PROGRESS, details.job_id), or another maintenance task holds the lock (MAINTENANCE_IN_PROGRESS)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/admin/photos/cleanup:
    get:
      tags: [admin]
      summary: Scan for clearly broken photos
      description: |
        Reports indexed photos whose original is missing or zero bytes. A photo
        is only reported missing when its library root (users/{id} or
        shared/{id}) is present, so an unmounted volume never looks like mass
        deletion; such photos and unreadable entries are counted in skipped.
        The scan is synchronous and holds the maintenance lock while it runs.
      operationId: scanBrokenPhotos
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Scan result; nothing is deleted
          content: { application/json: { schema: { $ref: '#/components/schemas/BrokenPhotoScanResult' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminRequired' }
        '409': { $ref: '#/components/responses/MaintenanceConflict' }
    post:
      tags: [admin]
      summary: Permanently delete confirmed broken photos
      description: |
        Deletes only the IDs the administrator confirmed from a previous scan,
        and only if each is still clearly broken now. IDs that are no longer
        broken or no longer exist are reported in failures and left untouched.
      operationId: cleanupBrokenPhotos
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrokenPhotoCleanupRequest' }
      responses:
        '200':
          description: Cleanup result
          content: { application/json: { schema: { $ref: '#/components/schemas/BrokenPhotoCleanupResult' } } }
        '400':
          description: Invalid JSON (INVALID_REQUEST)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminOrCSRFForbidden' }
        '409': { $ref: '#/components/responses/MaintenanceConflict' }
        '422':
          description: confirm was not true or ids was empty (CONFIRMATION_REQUIRED), or more than 10000 ids (BATCH_TOO_LARGE)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/v1/admin/imports:
    post:
      tags: [admin]
      summary: Move loose files from the photo root into a user's library
      description: |
        source_path is the one client-supplied filesystem location in the API.
        It must be relative to the photo root, must not contain "..", absolute
        paths, backslashes or NUL, must not start with users/, shared/ or a
        dot directory, and no component may be a symlink. "." imports root
        files outside users/, shared/ and hidden directories. Files move to
        users/{user_id}/Imported, keeping subfolders or, with
        organize_by_date, as Imported/YYYY/MM/DD using the timeline capture
        date. Existing destination files are never overwritten. Moved files
        are indexed by a follow-up rescan before the job completes.
      operationId: startImport
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CSRFHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportRequest' }
      responses:
        '202':
          description: Import queued
          content: { application/json: { schema: { $ref: '#/components/schemas/ImportJob' } } }
        '400':
          description: Invalid JSON (INVALID_JSON) or rejected source directory (INVALID_IMPORT_SOURCE)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AdminOrCSRFForbidden' }
        '404':
          description: Target user does not exist, is disabled or deleted (USER_NOT_FOUND)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '409':
          description: An import is already running (IMPORT_IN_PROGRESS), or another maintenance task holds the lock (MAINTENANCE_IN_PROGRESS)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  /api/v1/admin/people:
    get:
      tags: [admin]
      summary: List people with valid photos
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: Page of people
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/FacePerson' }
                  next_cursor: { type: string }
...
  /api/v1/trash/photos:
    get:
      tags: [photos]
      summary: List the current owner's trash, or all owners for administrators
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all], default: mine } }
        - { name: cursor, in: query, schema: { type: string }, description: Signed cursor bound to principal and scope }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Private non-cacheable page
          content: { application/json: { schema: { $ref: '#/components/schemas/TrashPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/WriteForbidden' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: 77photo_session
      description: HttpOnly server-side session cookie; never store in localStorage.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque short-lived mobile access token. Store it only in platform secure storage.
  headers:
    SessionCookie:
      schema: { type: string }
      description: HttpOnly session cookie. Secure is set in production.
...
  schemas:
...
    SetupAdminRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 256, format: password }
    SetupStatusResponse:
      type: object
      required: [required]
      properties:
        required: { type: boolean }
...
    AuthResponse:
      type: object
      required: [user, csrf_token]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrf_token: { type: string }
...